+++ examples/binary/main.tf
@@ -7,7 +7,6 @@ module "binary_string" {
   source        = "./../../"
   name          = "example-binary-secret"
   description   = "a binary secret example"
-  kms_key_id    = data.aws_kms_alias.secretsmanager.target_key_arn
   secret_policy = local.policy
   secrets = {
     secret1 = {