+++ examples/complete/data.tf
@@ -2,8 +2,16 @@ data "aws_caller_identity" "current" {}
 
 data "aws_partition" "current" {}
 
-data "aws_kms_alias" "secretsmanager" {
-  name = "alias/aws/secretsmanager"
+data "aws_region" "current" {}
+
+data "aws_availability_zones" "available" {
+  state = "available"
+}
+
+data "archive_file" "lambda" {
+  type        = "zip"
+  source_dir  = "mysql-lambda"
+  output_path = local.filename
 }
 
 data "aws_iam_policy_document" "assume_role_policy" {
@@ -14,11 +22,48 @@ data "aws_iam_policy_document" "assume_role_policy" {
       type        = "Service"
       identifiers = ["lambda.amazonaws.com"]
     }
+
+    effect = "Allow"
   }
 }
 
-data "archive_file" "lambda" {
-  type        = "zip"
-  source_dir  = "secrets_manager_rotation"
-  output_path = local.filename
+data "aws_iam_policy_document" "mysql_lambda_policy" {
+  #checkov:skip=CKV_AWS_111: "Ensure IAM policies does not allow write access without constraints"
+  statement {
+    actions = [
+      "ec2:CreateNetworkInterface",
+      "ec2:DeleteNetworkInterface",
+      "ec2:DescribeNetworkInterfaces",
+      "ec2:DetachNetworkInterface",
+    ]
+    resources = ["*", ]
+  }
+  statement {
+    actions = [
+      "secretsmanager:DescribeSecret",
+      "secretsmanager:GetSecretValue",
+      "secretsmanager:PutSecretValue",
+      "secretsmanager:UpdateSecretVersionStage",
+    ]
+    resources = [
+      "arn:${data.aws_partition.current.partition}:secretsmanager:${data.aws_region.current.name}:${data.aws_caller_identity.current.account_id}:secret:*",
+    ]
+  }
+  statement {
+    actions   = ["secretsmanager:GetRandomPassword"]
+    resources = ["*", ]
+  }
+}
+
+data "aws_iam_policy_document" "monitoring" {
+  statement {
+    actions = [
+      "sts:AssumeRole",
+    ]
+
+    principals {
+      type        = "Service"
+      identifiers = ["monitoring.rds.amazonaws.com"]
+    }
+  }
 }