+++ examples/complete/main.tf
@@ -1,59 +1,117 @@
-resource "aws_iam_role" "lambda" {
-  name               = "${local.name}-iam-role-lambda"
-  assume_role_policy = data.aws_iam_policy_document.assume_role_policy.json
-}
-resource "aws_lambda_function" "sample_mysql" {
-  #checkov:skip=CKV_AWS_117: "Ensure that AWS Lambda function is configured inside a VPC"
-  #checkov:skip=CKV_AWS_116: "Ensure that AWS Lambda function is configured for a Dead Letter Queue(DLQ)"
-  filename                       = local.filename
-  function_name                  = "${local.name}-rotation"
-  handler                        = "secrets_manager_rotation.lambda_handler"
-  runtime                        = "python3.7"
-  role                           = aws_iam_role.lambda.arn
-  description                    = "AWS SecretsManager secret rotation for RDS MySQL using single user."
-  source_code_hash               = data.archive_file.lambda.output_base64sha256
-  reserved_concurrent_executions = 0
-  tracing_config {
-    mode = "Active"
-  }
-  depends_on = [
-    data.archive_file.lambda
-  ]
-}
-
-resource "aws_lambda_permission" "default" {
-  function_name = aws_lambda_function.sample_mysql.function_name
-  statement_id  = "AllowExecutionSecretManager"
-  action        = "lambda:InvokeFunction"
-  principal     = "secretsmanager.amazonaws.com"
-}
-
-resource "random_password" "rds_password" {
-  length  = 16
-  special = false
-}
 
 module "secret_rotation" {
   source                                = "./../../"
-  name                                  = "${local.name}-rotation"
-  description                           = "Example complete secet with rotation"
-  kms_key_id                            = data.aws_kms_alias.secretsmanager.target_key_arn
+  name                                  = local.name
+  description                           = "Example complete secret with rotation"
   enable_secretsmanager_secret_rotation = true
   secret_policy                         = local.policy
-  rotation_lambda_arn                   = aws_lambda_function.sample_mysql.arn
+  automatically_after_days              = 7
+  rotation_lambda_arn                   = aws_lambda_function.mysql.arn
   secrets = {
     secret1 = {
       secret_string = jsonencode(
         {
-          username = "admin"
-          password = random_password.rds_password.result
-          engine   = "mysql"
-          port     = "3306"
+          engine   = aws_db_instance.mysql.engine
+          host     = aws_db_instance.mysql.address
+          username = aws_db_instance.mysql.username
+          password = random_password.mysql_password.result
+          dbname   = aws_db_instance.mysql.db_name
+          port     = aws_db_instance.mysql.port
       })
     }
   }
   tags = {
     environment        = "examples"
     "user::CostCenter" = "terraform-registry"
+  }
+}
+
+module "rotation_vpc" {
+  source               = "git::https://github.com/boldlink/terraform-aws-vpc.git?ref=2.0.3"
+  name                 = "${local.name}-vpc"
+  account              = data.aws_caller_identity.current.account_id
+  region               = data.aws_region.current.name
+  tag_env              = local.tag_env
+  cidr_block           = local.cidr_block
+  enable_dns_support   = true
+  enable_dns_hostnames = true
+  private_subnets      = local.rotation_subnets
+  availability_zones   = local.azs
+}
+
+resource "aws_vpc_endpoint" "rotation_vpc" {
+  vpc_id            = module.rotation_vpc.id
+  service_name      = "com.amazonaws.${data.aws_region.current.name}.secretsmanager"
+  vpc_endpoint_type = "Interface"
+  subnet_ids        = flatten(module.rotation_vpc.private_subnet_id)
+
+  security_group_ids = [
+    aws_security_group.mysql.id,
+  ]
+
+  private_dns_enabled = true
+}
+
+resource "aws_security_group" "mysql" {
+  name        = "${local.name}-security-group"
+  description = "Allow inbound traffic"
+  vpc_id      = module.rotation_vpc.id
+
+  ingress {
+    cidr_blocks     = [local.cidr_block]
+    description     = "mysql ingress rule"
+    from_port       = 0
+    prefix_list_ids = []
+    protocol        = "-1"
+    security_groups = [aws_security_group.lambda.id]
+    self            = false
+    to_port         = 0
+  }
+  egress {
+    cidr_blocks     = [local.cidr_block]
+    description     = "mysql egress rule"
+    from_port       = 0
+    prefix_list_ids = []
+    protocol        = "-1"
+    security_groups = []
+    self            = false
+    to_port         = 0
+  }
+
+  tags = {
+    Name = local.name
+  }
+  lifecycle {
+    create_before_destroy = true
+  }
+}
+
+resource "aws_security_group" "lambda" {
+  name        = "${local.name}-lambda-security-group"
+  description = "Allow inbound traffic"
+  vpc_id      = module.rotation_vpc.id
+
+  ingress {
+    cidr_blocks     = [local.cidr_block]
+    description     = "lambda function ingress rule"
+    from_port       = 0
+    prefix_list_ids = []
+    protocol        = "-1"
+    security_groups = []
+    self            = false
+    to_port         = 0
+  }
+  egress {
+    cidr_blocks     = [local.cidr_block]
+    description     = ""
+    from_port       = 0
+    prefix_list_ids = []
+    protocol        = "-1"
+    security_groups = []
+    self            = false
+    to_port         = 0
+  }
+  lifecycle {
+    create_before_destroy = true
   }
 }