+++ examples/minimum/data.tf
@@ -1,4 +1,0 @@
-
-data "aws_kms_alias" "secretsmanager" {
-  name = "alias/aws/secretsmanager"
-}